+++ utils/CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Changed
 
+## [v2.1.0] 2025-11-XX
+### Added
+- [PKG-18] Set minimum go version as 1.23
+- [PKG-18] Add struct OrderedMap[K]V;
+- [PKG-28] Add struct ObjectPool[T];
+### Fixed
+### Changed
+
 ## [v2.0.0] 2025-10-23
 ### Added
 - Set minimum go version as 1.22